# api/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: kusk-gateway-api
  description: "This is the Kusk Gateway Management API"
servers:
  - url: http://localhost:8080
    description: "My local endpoint mockup"
...
paths:
...
  /staticroutes:
    get:
      tags:
        - "static routes"
      summary: "Get a list of static routes"
      description: "Returns a list of static routes"
      operationId: getStaticRoutes
      parameters:
        - name: namespace
          in: query
          schema:
            type: string
          required: false
          description: "optional filter on namespace"
      responses:
        200:
          description: "list of static routes"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaticRouteItem"
    post:
      tags:
        - "create new static route"
      summary: "create new static route"
      operationId: createStaticRoute
      requestBody:
        description: "static route content"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                namespace:
                  type: string
                envoyFleetName:
                  type: string
                envoyFleetNamespace:
                  type: string
                openapi:
                  type: string

      responses:
        200:
          description: "created static route"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/StaticRouteItem"
        400:
          description: "The content of the API is malformed"
          content:
            application/json:
              schema:
                type: string
  /staticroutes/{namespace}/{name}:
    get:
      tags:
        - "static routes"
      summary: "Get details for a single static route"
      description: "Returns an object containing info about the static route corresponding to the namespace and name"
      operationId: getStaticRoute
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
          description: "the namespace of the static route"
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: "the name of the static route"
      responses:
        200:
          description: "get static route details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticRouteItem"
        404:
          description: "Static Route not found by namespace-name combination"
    put:
      tags:
        - "static routes"
      summary: "Update a static route by namespace and name"
      operationId: "updateStaticRoute"
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
            default: "default"
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: "static route content"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                namespace:
                  type: string
                envoyFleetName:
                  type: string
                envoyFleetNamespace:
                  type: string
                openapi:
                  type: string
      responses:
        201:
          description: "static route updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticRouteItem"
        400:
          description: "The content of the static route is malformed"
          content:
            application/json:
              schema:
                type: string
        404:
          description: "static route not found"
    delete:
      tags:
        - "static Route"
      summary: "Delete a StaticRoute by namespace and name"
      operationId: deleteStaticRoute
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
            default: "default"
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: "staticRoute item deleted"
        404:
          description: "staticRoute item not found"
  /staticroutes/{namespace}/{name}/crd:
    get:
      tags:
        - "static routes"
      summary: "Get static route CRD"
      operationId: getStaticRouteCRD
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
            default: "default"
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Static route CRD"
          content:
            application/json:
              schema:
                type: object
        404:
          description: "Static route CRD not found"
...
components:
  schemas:
...
    StaticRouteItem:
      type: object
      required:
        - name
        - namespace
      properties:
        name:
          type: string
        namespace:
          type: string
        envoyFleetName:
          type: string
        envoyFleetNamespace:
          type: string
...